refactor: move and profiling redis application
