+++ chat-server/src/main/java/com/carrot/chat/redis/application/converter/MessageConverter.java
@@ -1,12 +1,14 @@
-package com.carrot.chat.support.converter;
+package com.carrot.chat.redis.application.converter;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+import org.springframework.context.annotation.Profile;
 import org.springframework.data.redis.connection.Message;
 
 import java.io.IOException;
 
+@Profile("redis-pub-sub")
 public class MessageConverter {
     private MessageConverter() {
