+++ chat-server/src/main/java/com/carrot/chat/redis/config/EmbeddedRedisConfig.java
@@ -1,12 +1,14 @@
-package com.carrot.chat.queue.config;
+package com.carrot.chat.redis.config;
 
 import jakarta.annotation.PostConstruct;
 import jakarta.annotation.PreDestroy;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
 import redis.embedded.RedisServer;
 
 @Configuration
+@Profile("redis-pub-sub")
 public class EmbeddedRedisConfig {
 
     @Value("${spring.redis.property.port}")
