+++ chat-server/src/main/java/com/carrot/chat/redis/config/RedisConfig.java
@@ -1,11 +1,12 @@
-package com.carrot.chat.queue.config;
+package com.carrot.chat.redis.config;
 
-import com.carrot.chat.queue.ui.MessageObject;
+import com.carrot.chat.redis.ui.MessageObject;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
 import org.springframework.data.redis.connection.RedisConnectionFactory;
 import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
 import org.springframework.data.redis.core.RedisTemplate;
@@ -14,6 +15,7 @@ import org.springframework.data.redis.serializer.Jackson2JsonRedisSerializer;
 import org.springframework.data.redis.serializer.StringRedisSerializer;
 
 @Configuration
+@Profile("redis-pub-sub")
 public class RedisConfig {
 
     @Bean
