+++ chat-server/src/main/java/com/carrot/chat/redis/ui/ChatController.java
@@ -1,6 +1,7 @@
-package com.carrot.chat.queue.ui;
+package com.carrot.chat.redis.ui;
 
-import com.carrot.chat.queue.application.grpc.ChatGrpcClient;
+import com.carrot.chat.support.client.ChatGrpcClient;
+import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -8,7 +9,8 @@ import org.springframework.web.bind.annotation.PathVariable;
 
 import java.util.List;
 
-@Controller
+@Controller("/redis")
+@Profile("redis-pub-sub")
 public class ChatController {
 
     private final ChatGrpcClient chatGrpcClient;
