+++ chat-server/src/main/java/com/carrot/chat/support/client/AuctionGrpcClient.java
@@ -1,4 +1,4 @@
-package com.carrot.chat.queue.application.grpc;
+package com.carrot.chat.support.client;
 
 import auctions.AuctionItemServiceGrpc;
 import auctions.Auctions;
