+++ chat-server/src/main/java/com/carrot/chat/support/config/GrpcConfig.java
@@ -1,4 +1,4 @@
-package com.carrot.chat.queue.config;
+package com.carrot.chat.support.config;
 
 
 import auctions.AuctionItemServiceGrpc;
