+++ chat-server/src/main/java/com/carrot/chat/websocket/chatmessage/application/MessageListenerFactory.java
@@ -1,12 +1,13 @@
 package com.carrot.chat.websocket.chatmessage.application;
 
-import com.carrot.chat.queue.application.Subscriber;
-import com.carrot.chat.queue.config.CustomMessageListenerAdapter;
+import com.carrot.chat.redis.application.Subscriber;
+import com.carrot.chat.redis.config.CustomMessageListenerAdapter;
+import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
 
 @Component
+@Profile("redis-pub-sub")
 public class MessageListenerFactory {
-
 
     public CustomMessageListenerAdapter create(String name) {
         return new CustomMessageListenerAdapter(new Subscriber(name));
