+++ chat-server/src/main/java/com/carrot/chat/websocket/config/MongoConfig.java
@@ -1,4 +1,4 @@
-package com.carrot.chat.websocket.common;
+package com.carrot.chat.websocket.config;
 
 import com.mongodb.reactivestreams.client.MongoClient;
 import com.mongodb.reactivestreams.client.MongoClients;
