+++ chat-server/src/main/resources/application-redis-pub-sub.properties
@@ -1,0 +1,16 @@
+spring.application.name=CarrotRedisServer
+spring.datasource.password=password
+spring.redis.property.port=6378
+
+logging.level.root=info
+spring.main.banner-mode=off
+
+spring.main.web-application-type=reactive
+
+server.port=8001
+
+logging.level.org.springframework.data.mongodb.core.ReactiveMongoTemplate=DEBUG
+
+spring.data.jpa.repositories.enabled=false
+spring.autoconfigure.exclude=org.springframework.boot.autoconfigure.mongo.MongoReactiveAutoConfiguration
+
